Undid the abstract class change because Xunit doesn't like it
