+++ tests/Microsoft.DotNet.Docker.Tests/AspnetCompositeImageTests.cs
@@ -1,4 +1,4 @@
-// Licensed to the .NET Foundation under one or more agreements.
+﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
@@ -8,87 +8,115 @@ using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Microsoft.DotNet.Docker.Tests;
-
-[Trait("Category", "aspnet-composite")]
-public class AspnetCompositeImageTests : CommonAspnetImageTests
+namespace Microsoft.DotNet.Docker.Tests
 {
-    public AspnetCompositeImageTests(ITestOutputHelper outputHelper)
-        : base(outputHelper)
+    [Trait("Category", "aspnet-composite")]
+    public class AspnetCompositeImageTests : CommonRuntimeImageTests
     {
-    }
-
-    protected override DotNetImageType ImageType => DotNetImageType.Aspnet_Composite;
-
-    public static IEnumerable<object[]> GetImageData() => GetImageData(DotNetImageType.Aspnet_Composite);
-
-    [DotNetTheory]
-    [MemberData(nameof(GetImageData))]
-    public override async Task VerifyAppScenario(ProductImageData imageData)
-    {
-        if (imageData.OSTag != OS.Alpine317Composite)
+        public AspnetCompositeImageTests(ITestOutputHelper outputHelper)
+            : base(outputHelper)
         {
-            OutputHelper.WriteLine("Skipping because currently only Alpine composite"
-                                 + " images are supported.");
-            return ;
         }
 
-        ImageScenarioVerifier verifier = new ImageScenarioVerifier(
-                                            imageData: imageData,
-                                            dockerHelper: DockerHelper,
-                                            outputHelper: OutputHelper,
-                                            isWeb: true);
-        await verifier.Execute();
-    }
+        protected override DotNetImageType ImageType => DotNetImageType.Aspnet_Composite;
 
-    [DotNetTheory]
-    [MemberData(nameof(GetImageData))]
-    public override void VerifyEnvironmentVariables(ProductImageData imageData)
-    {
-        List<EnvironmentVariableInfo> variables = new List<EnvironmentVariableInfo>
+        public static IEnumerable<object[]> GetImageData() => GetImageData(DotNetImageType.Aspnet_Composite);
+
+        [DotNetTheory]
+        [MemberData(nameof(GetImageData))]
+        public async Task VerifyAppScenario(ProductImageData imageData)
         {
-            RuntimeImageTests.GetRuntimeVersionVariableInfo(imageData, DockerHelper)
-        };
+            if (imageData.OSTag != OS.Alpine317Composite)
+            {
+                OutputHelper.WriteLine("Skipping because currently only Alpine composite"
+                                     + " images are supported.");
+                return ;
+            }
 
-        EnvironmentVariableInfo compositeVersionVariableInfo = GetAspnetCompositeVersionVariableInfo(
-                imageData,
-                DockerHelper);
-
-        if (compositeVersionVariableInfo != null)
-        {
-            variables.Add(compositeVersionVariableInfo);
+            ImageScenarioVerifier verifier = new ImageScenarioVerifier(
+                                                imageData: imageData,
+                                                dockerHelper: DockerHelper,
+                                                outputHelper: OutputHelper,
+                                                isWeb: true);
+            await verifier.Execute();
         }
 
-        base.VerifyCommonEnvironmentVariables(imageData, variables);
-    }
+        [DotNetTheory]
+        [MemberData(nameof(GetImageData))]
+        public void VerifyEnvironmentVariables(ProductImageData imageData)
+        {
+            List<EnvironmentVariableInfo> variables = new List<EnvironmentVariableInfo>
+            {
+                RuntimeImageTests.GetRuntimeVersionVariableInfo(imageData, DockerHelper)
+            };
 
-    [DotNetTheory]
-    [MemberData(nameof(GetImageData))]
-    public override void VerifyPackageInstallation(ProductImageData imageData)
-    {
-        VerifyExpectedInstalledRpmPackages(
+            EnvironmentVariableInfo compositeVersionVariableInfo = GetAspnetCompositeVersionVariableInfo(
+                    imageData,
+                    DockerHelper);
+
+            if (compositeVersionVariableInfo != null)
+            {
+                variables.Add(compositeVersionVariableInfo);
+            }
+
+            base.VerifyCommonEnvironmentVariables(imageData, variables);
+        }
+
+        [DotNetTheory]
+        [MemberData(nameof(GetImageData))]
+        public void VerifyPackageInstallation(ProductImageData imageData)
+        {
+            VerifyExpectedInstalledRpmPackages(
                 imageData,
                 GetExpectedRpmPackagesInstalled(imageData)
-                .Concat(RuntimeDepsImageTests.GetExpectedRpmPackagesInstalled(imageData)));
-    }
+                .Concat(RuntimeImageTests.GetExpectedRpmPackagesInstalled(imageData)));
+        }
 
-    public static EnvironmentVariableInfo GetAspnetCompositeVersionVariableInfo(
-            ProductImageData imageData,
-            DockerHelper dockerHelper)
-    {
-        string version = imageData.GetProductVersion(DotNetImageType.Aspnet_Composite,
-                                                     dockerHelper);
-
-        return new EnvironmentVariableInfo("ASPNET_VERSION", version)
+        [LinuxImageTheory]
+        [MemberData(nameof(GetImageData))]
+        public void VerifyInsecureFiles(ProductImageData imageData)
         {
-            IsProductVersion = true
-        };
-    }
+            base.VerifyCommonInsecureFiles(imageData);
+        }
 
-    internal static string[] GetExpectedRpmPackagesInstalled(ProductImageData imageData) =>
-        new string[]
+        [LinuxImageTheory]
+        [MemberData(nameof(GetImageData))]
+        public void VerifyShellNotInstalledForDistroless(ProductImageData imageData)
         {
-            $"aspnetcore-runtime-composite-{imageData.VersionString}"
-        };
+            base.VerifyCommonShellNotInstalledForDistroless(imageData);
+        }
+
+        [DotNetTheory]
+        [MemberData(nameof(GetImageData))]
+        public void VerifyNoSasToken(ProductImageData imageData)
+        {
+            base.VerifyCommonNoSasToken(imageData);
+        }
+
+        [DotNetTheory]
+        [MemberData(nameof(GetImageData))]
+        public void VerifyDefaultUser(ProductImageData imageData)
+        {
+            base.VerifyCommonDefaultUser(imageData);
+        }
+
+        public static EnvironmentVariableInfo GetAspnetCompositeVersionVariableInfo(
+                ProductImageData imageData,
+                DockerHelper dockerHelper)
+        {
+            string version = imageData.GetProductVersion(DotNetImageType.Aspnet_Composite,
+                                                         dockerHelper);
+
+            return new EnvironmentVariableInfo("ASPNET_VERSION", version)
+            {
+                IsProductVersion = true
+            };
+        }
+
+        internal static string[] GetExpectedRpmPackagesInstalled(ProductImageData imageData) =>
+            new string[]
+            {
+                $"aspnetcore-runtime-composite-{imageData.VersionString}"
+            };
+    }
 }
-
